pyproject.toml
[project]
name = "at-mind-group-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datapizza-ai>=0.0.2",
    "datapizza-ai-clients-openai-like",
    "fastapi",
    "uvicorn",
]
